Fix linting, move ContextMixin into _meta_mixins
