+++ src/hera/workflows/steps.py
@@ -6,10 +6,9 @@ for more on Steps.
 
 from typing import Any, List, Optional, Union
 
-from hera.workflows._meta_mixins import CallableTemplateMixin
+from hera.workflows._meta_mixins import CallableTemplateMixin, ContextMixin
 from hera.workflows._mixins import (
     ArgumentsMixin,
-    ContextMixin,
     IOMixin,
     ItemMixin,
     ParameterMixin,
